Moved coordinator classes to its own file
